refactor: always load main environment file
